requirements.txt
genkit
genkit-plugin-google-genai
